Bundle libpsgplay.js in psgplay.min.js instead of lazy-loading
Fixes #29

`onRuntimeInitialized` only fires once across all emscripten modules, so we can't use it to detect when libpsgplay is loaded if that mechanism has already been spent on libopenmpt. However, we don't really need that mechanism - we only care about lazy-loading the wasm file, not libpsgplay.js (which is only 26kb), and current emscripten provides the MODULARIZE option for doing that.

As a bonus, we no longer need the pathToLibPSGPlay opton.